refactor: organize hook types into `src/hooks`
